Separate diagtree into its own package
