+++ internal/cmd/compare.go
@@ -11,10 +11,10 @@ import (
 	"strings"
 
 	"github.com/pulumi/pulumi/pkg/v3/codegen/schema"
-	"github.com/pulumi/pulumi/sdk/v3/go/common/util/contract"
 	"github.com/spf13/cobra"
 
 	"github.com/pulumi/schema-tools/internal/pkg"
+	"github.com/pulumi/schema-tools/internal/util/diagtree"
 	"github.com/pulumi/schema-tools/internal/util/set"
 )
 
@@ -81,146 +81,8 @@ func compare(provider string, oldCommit string, newCommit string) error {
 	return nil
 }
 
-type message struct {
-	Title       string
-	Description string
-	Severity    Severity
-	Subfields   []*message
-
-	doDisplay bool
-	parent    *message
-}
-
-func (m *message) subfield(name string) *message {
-	contract.Assertf(name != "", "we cannot display an empty name")
-	for _, v := range m.Subfields {
-		if v.Title == name {
-			return v
-		}
-	}
-	v := &message{
-		Title:  name,
-		parent: m,
-	}
-	m.Subfields = append(m.Subfields, v)
-	return v
-}
-
-func (m *message) label(name string) *message {
-	return m.subfield(name)
-}
-
-func (m *message) value(value string) *message {
-	return m.subfield(fmt.Sprintf("%q", value))
-}
-
-func (m *message) levelPrefix(level int) string {
-	switch level {
-	case 0:
-		return "###"
-	case 1:
-		return "####"
-	}
-	if level < 0 {
-		return ""
-	}
-	return strings.Repeat("  ", (level-2)*2) + "-"
-}
-
-func (m *message) display(out io.Writer, level int, prefix bool, max int) int {
-	if m == nil || !m.doDisplay || max <= 0 {
-		// Nothing to display
-		return 0
-	}
-
-	var displayed int
-	var display string
-	if m.Title != "" {
-		if prefix {
-			display = fmt.Sprintf("%s %s ",
-				m.levelPrefix(level),
-				m.severity())
-		}
-		display += m.Title + ": "
-		if m.Description != "" {
-			displayed += 1
-			display += m.Description
-		}
-
-		out.Write([]byte(display))
-	}
-
-	if level > 1 && m.Severity == None {
-		if s := m.uniqueSuccessor(); s != nil {
-			return s.display(out, level, false, max-displayed) + displayed
-		}
-	}
-
-	out.Write([]byte{'\n'})
-
-	order := make([]int, len(m.Subfields))
-	for i := range order {
-		order[i] = i
-	}
-	if level > 0 {
-		// Obtain an ordering on the subfields without mutating `.Subfields`.
-		sort.Slice(order, func(i, j int) bool {
-			return m.Subfields[order[i]].Title < m.Subfields[order[j]].Title
-		})
-	}
-
-	for _, f := range order {
-		n := m.Subfields[f].display(out, level+1, true, max-displayed)
-		displayed += n
-	}
-	return displayed
-}
-
-func (m *message) uniqueSuccessor() *message {
-	var us *message
-	for _, s := range m.Subfields {
-		if !s.doDisplay {
-			continue
-		}
-		if us != nil {
-			return nil
-		}
-		us = s
-	}
-	return us
-}
-
-func (m *message) severity() Severity {
-	for m != nil {
-		s := m.uniqueSuccessor()
-		if s == nil {
-			return m.Severity
-		}
-		m = s
-	}
-	return None
-}
-
-type Severity string
-
-const (
-	None   Severity = ""
-	Info   Severity = "`🟢`"
-	Warn   Severity = "`🟡`"
-	Danger Severity = "`🔴`"
-)
-
-func (m *message) SetDescription(level Severity, msg string, a ...any) {
-	for v := m.parent; v != nil && !v.doDisplay; v = v.parent {
-		v.doDisplay = true
-	}
-	m.doDisplay = true
-	m.Description = fmt.Sprintf(msg, a...)
-	m.Severity = level
-}
-
-func breakingChanges(oldSchema, newSchema schema.PackageSpec) *message {
-	msg := &message{Title: ""}
+func breakingChanges(oldSchema, newSchema schema.PackageSpec) *diagtree.Node {
+	msg := &diagtree.Node{Title: ""}
 
 	changedToRequired := func(kind, name string) string {
 		return fmt.Sprintf("%s %q has changed to Required", kind, name)
@@ -230,18 +92,18 @@ func breakingChanges(oldSchema, newSchema schema.PackageSpec) *message {
 	}
 
 	for resName, res := range oldSchema.Resources {
-		msg := msg.label("Resources").value(resName)
+		msg := msg.Label("Resources").Value(resName)
 		newRes, ok := newSchema.Resources[resName]
 		if !ok {
-			msg.SetDescription(Danger, "missing")
+			msg.SetDescription(diagtree.Danger, "missing")
 			continue
 		}
 
 		for propName, prop := range res.InputProperties {
-			msg := msg.label("inputs").value(propName)
+			msg := msg.Label("inputs").Value(propName)
 			newProp, ok := newRes.InputProperties[propName]
 			if !ok {
-				msg.SetDescription(Warn, "missing")
+				msg.SetDescription(diagtree.Warn, "missing")
 				continue
 			}
 
@@ -249,7 +111,7 @@ func breakingChanges(oldSchema, newSchema schema.PackageSpec) *message {
 		}
 
 		for propName, prop := range res.Properties {
-			msg := msg.label("properties").value(propName)
+			msg := msg.Label("properties").Value(propName)
 			newProp, ok := newRes.Properties[propName]
 			if !ok {
 				msg.SetDescription("missing output %q", propName)
@@ -261,15 +123,15 @@ func breakingChanges(oldSchema, newSchema schema.PackageSpec) *message {
 
 		oldRequiredInputs := set.FromSlice(res.RequiredInputs)
 		for _, input := range newRes.RequiredInputs {
-			msg := msg.label("required inputs").value(input)
+			msg := msg.Label("required inputs").Value(input)
 			if !oldRequiredInputs.Has(input) {
-				msg.SetDescription(Info, changedToRequired("input", input))
+				msg.SetDescription(diagtree.Info, changedToRequired("input", input))
 			}
 		}
 
 		newRequiredProperties := set.FromSlice(newRes.Required)
 		for _, prop := range res.Required {
-			msg := msg.label("required").value(prop)
+			msg := msg.Label("required").Value(prop)
 			// It is a breaking change to move an output property from
 			// required to optional.
 			//
@@ -277,23 +139,23 @@ func breakingChanges(oldSchema, newSchema schema.PackageSpec) *message {
 			// already warned on, so we don't need to warn here.
 			_, stillExists := newRes.Properties[prop]
 			if !newRequiredProperties.Has(prop) && stillExists {
-				msg.SetDescription(Info, changedToOptional("property", prop))
+				msg.SetDescription(diagtree.Info, changedToOptional("property", prop))
 			}
 		}
 	}
 
 	for funcName, f := range oldSchema.Functions {
-		msg := msg.label("Functions").value(funcName)
+		msg := msg.Label("Functions").Value(funcName)
 		newFunc, ok := newSchema.Functions[funcName]
 		if !ok {
-			msg.SetDescription(Danger, "missing")
+			msg.SetDescription(diagtree.Danger, "missing")
 			continue
 		}
 
 		if f.Inputs != nil {
-			msg := msg.label("inputs")
+			msg := msg.Label("inputs")
 			for propName, prop := range f.Inputs.Properties {
-				msg := msg.label("properties").value(propName)
+				msg := msg.Label("properties").Value(propName)
 				if newFunc.Inputs == nil {
 					msg.SetDescription("missing input %q", propName)
 					continue
@@ -309,29 +171,29 @@ func breakingChanges(oldSchema, newSchema schema.PackageSpec) *message {
 			}
 
 			if newFunc.Inputs != nil {
-				msg := msg.label("required")
+				msg := msg.Label("required")
 				oldRequired := set.FromSlice(f.Inputs.Required)
 				for _, req := range newFunc.Inputs.Required {
-					msg.value(req)
+					msg.Value(req)
 					if !oldRequired.Has(req) {
-						msg.SetDescription(Info, changedToRequired("input", req))
+						msg.SetDescription(diagtree.Info, changedToRequired("input", req))
 					}
 				}
 			}
 		}
 
 		if f.Outputs != nil {
-			msg := msg.label("outputs")
+			msg := msg.Label("outputs")
 			for propName, prop := range f.Outputs.Properties {
-				msg := msg.label("properties").value(propName)
+				msg := msg.Label("properties").Value(propName)
 				if newFunc.Outputs == nil {
-					msg.SetDescription(Warn, "missing output")
+					msg.SetDescription(diagtree.Warn, "missing output")
 					continue
 				}
 
 				newProp, ok := newFunc.Outputs.Properties[propName]
 				if !ok {
-					msg.SetDescription(Warn, "missing output")
+					msg.SetDescription(diagtree.Warn, "missing output")
 					continue
 				}
 
@@ -345,25 +207,26 @@ func breakingChanges(oldSchema, newSchema schema.PackageSpec) *message {
 			for _, req := range f.Outputs.Required {
 				_, stillExists := f.Outputs.Properties[req]
 				if !newRequired.Has(req) && stillExists {
-					msg.label("required").value(req).SetDescription(Info, changedToOptional("property", req))
+					msg.Label("required").Value(req).SetDescription(
+						diagtree.Info, changedToOptional("property", req))
 				}
 			}
 		}
 	}
 
 	for typName, typ := range oldSchema.Types {
-		msg := msg.label("Types").value(typName)
+		msg := msg.Label("Types").Value(typName)
 		newTyp, ok := newSchema.Types[typName]
 		if !ok {
-			msg.SetDescription(Danger, "missing")
+			msg.SetDescription(diagtree.Danger, "missing")
 			continue
 		}
 
 		for propName, prop := range typ.Properties {
-			msg := msg.label("properties").value(propName)
+			msg := msg.Label("properties").Value(propName)
 			newProp, ok := newTyp.Properties[propName]
 			if !ok {
-				msg.SetDescription(Warn, "missing")
+				msg.SetDescription(diagtree.Warn, "missing")
 				continue
 			}
 
@@ -377,13 +240,15 @@ func breakingChanges(oldSchema, newSchema schema.PackageSpec) *message {
 		for _, r := range typ.Required {
 			_, stillExists := typ.Properties[r]
 			if !newRequired.Has(r) && stillExists {
-				msg.label("required").value(r).SetDescription(Info, changedToOptional("property", r))
+				msg.Label("required").Value(r).SetDescription(
+					diagtree.Info, changedToOptional("property", r))
 			}
 		}
 		required := set.FromSlice(typ.Required)
 		for _, r := range newTyp.Required {
 			if !required.Has(r) {
-				msg.label("required").value(r).SetDescription(Info, changedToRequired("property", r))
+				msg.Label("required").Value(r).SetDescription(
+					diagtree.Info, changedToRequired("property", r))
 			}
 		}
 	}
@@ -395,7 +260,7 @@ func compareSchemas(out io.Writer, provider string, oldSchema, newSchema schema.
 	fmt.Fprintf(out, "### Does the PR have any schema changes?\n\n")
 	violations := breakingChanges(oldSchema, newSchema)
 	displayedViolations := new(bytes.Buffer)
-	lenViolations := violations.display(displayedViolations, 0, true, 500)
+	lenViolations := violations.Display(displayedViolations, 500)
 	switch lenViolations {
 	case 0:
 		fmt.Fprintln(out, "Looking good! No breaking changes found.")
@@ -442,15 +307,15 @@ func compareSchemas(out io.Writer, provider string, oldSchema, newSchema schema.
 	}
 }
 
-func validateTypes(old *schema.TypeSpec, new *schema.TypeSpec, msg *message) {
+func validateTypes(old *schema.TypeSpec, new *schema.TypeSpec, msg *diagtree.Node) {
 	switch {
 	case old == nil && new == nil:
 		return
 	case old != nil && new == nil:
-		msg.SetDescription(Warn, "had %+v but now has no type", old)
+		msg.SetDescription(diagtree.Warn, "had %+v but now has no type", old)
 		return
 	case old == nil && new != nil:
-		msg.SetDescription(Warn, "had no type but now has %+v", new)
+		msg.SetDescription(diagtree.Warn, "had no type but now has %+v", new)
 		return
 	}
 
@@ -466,8 +331,8 @@ func validateTypes(old *schema.TypeSpec, new *schema.TypeSpec, msg *message) {
 		msg.SetDescription("type changed from %q to %q", oldType, newType)
 	}
 
-	validateTypes(old.Items, new.Items, msg.label("items"))
-	validateTypes(old.AdditionalProperties, new.AdditionalProperties, msg.label("additional properties"))
+	validateTypes(old.Items, new.Items, msg.Label("items"))
+	validateTypes(old.AdditionalProperties, new.AdditionalProperties, msg.Label("additional properties"))
 }
 
 func formatName(provider, s string) string {
